refact: remove upload.rs file
