+++ src/media.rs
@@ -1,7 +1,5 @@
 use std::str::FromStr;
 
-use serde::Deserialize;
-
 use crate::WeComError;
 
 pub enum MediaType {
@@ -42,45 +40,5 @@ impl MediaType {
         U: Into<String>,
     {
         format!("{}&type={}", base.into(), self.to_string())
-    }
-}
-
-#[derive(Debug, Deserialize)]
-pub struct UploadResp {
-    #[serde(rename = "errcode")]
-    pub err_code: i64,
-
-    #[serde(rename = "errmsg")]
-    pub err_msg: String,
-
-    #[serde(rename = "type", default)]
-    pub media_type: String,
-
-    #[serde(rename = "media_id", default)]
-    pub media_id: String,
-
-    #[serde(rename = "created_at", default)]
-    pub created_at: String,
-}
-
-impl Default for UploadResp {
-    fn default() -> Self {
-        UploadResp {
-            err_code: 0,
-            err_msg: String::from("success"),
-            media_type: MediaType::File.to_string(),
-            media_id: String::new(),
-            created_at: String::new(),
-        }
-    }
-}
-
-impl UploadResp {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    pub fn is_ok(&self) -> bool {
-        self.err_code.eq(&0)
     }
 }
